Removed the mock flavor for reasons of simplicity, Moved mock implementations into the test class (inner). Improved the structure of the build.gradle using version variables.
