+++ app/build.gradle
@@ -8,7 +8,7 @@ repositories {
     }
 }
 
-def downloadJarsToLibs(fileName, url) {
+static def downloadJarsToLibs(fileName, url) {
     def f = new File('app\\libs\\' + fileName)
     if (!f.exists()) {
         new URL(url).withInputStream { i -> f.withOutputStream { it << i } }
@@ -16,19 +16,19 @@ def downloadJarsToLibs(fileName, url) {
 }
 
 task downloadHostedJars() {
-    new File("app/libs").mkdirs();
+    new File("app/libs").mkdirs()
     def jars = ['zmanim-1.4.0alpha.jar': 'https://github.com/KosherJava/zmanim/raw/master/lib/zmanim-1.4.0alpha.jar'
     ]
     jars.each { fn, url -> downloadJarsToLibs(fn, url) }
 }
 
 android {
-    compileSdkVersion 25
+    compileSdkVersion 26
     buildToolsVersion "26.0.2"
     defaultConfig {
         applicationId "com.app.minyaneto_android"
         minSdkVersion 21
-        targetSdkVersion 25
+        targetSdkVersion 26
         versionCode 1
         versionName "1.0"
         multiDexEnabled true
@@ -38,15 +38,6 @@ android {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-
-    productFlavors {
-        mock {
-            applicationIdSuffix = ".mock"
-        }
-        prod {
-
         }
     }
 
@@ -80,8 +71,27 @@ dependencies {
     compile 'com.fasterxml.jackson.core:jackson-annotations:2.7.3'
     compile 'com.fasterxml.jackson.core:jackson-databind:2.7.3'
     compile files('libs/zmanim-1.4.0alpha.jar')
+    implementation "com.android.support:appcompat-v7:$rootProject.ext.supportLibVersion"
+    implementation "com.android.support.constraint:constraint-layout:1.0.2"
+    implementation "com.android.support:design:$rootProject.ext.supportLibVersion"
+    implementation "com.android.support:support-v4:$rootProject.ext.supportLibVersion"
+    implementation "com.android.support:cardview-v7:$rootProject.ext.supportLibVersion"
+    implementation "com.android.support:recyclerview-v7:$rootProject.ext.supportLibVersion"
+    implementation "com.google.android.gms:play-services-location:$rootProject.ext.playServicesVersion"
+    implementation "com.google.android.gms:play-services-maps:$rootProject.ext.playServicesVersion"
+    implementation "com.google.android.gms:play-services-places:$rootProject.ext.playServicesVersion"
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    implementation 'commons-io:commons-io:2.5'
+    implementation 'com.google.code.gson:gson:2.8.0'
+    implementation 'com.android.volley:volley:1.0.0'
+    implementation 'com.fasterxml.jackson.core:jackson-core:2.7.3'
+    implementation 'com.fasterxml.jackson.core:jackson-annotations:2.7.3'
+    implementation 'com.fasterxml.jackson.core:jackson-databind:2.7.3'
+    implementation files('libs/zmanim-1.4.0alpha.jar')
+    implementation "android.arch.lifecycle:extensions:1.0.0"
+    implementation "android.arch.lifecycle:extensions:1.0.0"
 
-    testCompile 'junit:junit:4.12'
-    testCompile "org.mockito:mockito-all:$rootProject.ext.mockitoVersion"
-    testCompile "org.hamcrest:hamcrest-all:$rootProject.ext.hamcrestVersion"
+    testImplementation 'junit:junit:4.12'
+    testImplementation "org.mockito:mockito-all:$rootProject.ext.mockitoVersion"
+    testImplementation "org.hamcrest:hamcrest-all:$rootProject.ext.hamcrestVersion"
 }
