+++ app/src/main/java/com/app/minyaneto_android/location/LocationProvider.java
@@ -1,11 +1,11 @@
 package com.app.minyaneto_android.location;
 
+import android.location.Location;
+
 import java.util.TimeZone;
 
 public interface LocationProvider {
     TimeZone getTimeZone();
 
-    double getLatitude();
-
-    double getLongitude();
+    Location getLocation();
 }
